+++ front/src/components/templates/components/MenuContent.styled.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const MenuContent = styled.nav`
+	display: flex;
+	justify-content: space-around;
+	align-items: center;
+	list-style: none;
+	width: ${(props) => props.style?.width || '700px'};
+	height: 41px;
+	font-weight: bold;
+`;
+
+export default MenuContent;
